# server/docs/swagger.yaml
basePath: /api
definitions:
...
  api.frpView:
    properties:
      agent_id:
        description: frps 行恒为 "server"（哨兵）；前端不展示
        type: string
      agent_name:
        type: string
      enabled:
        type: boolean
      kind:
        type: string
      proxies:
        items:
          $ref: '#/definitions/store.FrpTunnel'
        type: array
      server_addr:
        type: string
      server_port:
        type: integer
      status:
        $ref: '#/definitions/store.FrpStatus'
      token_set:
        type: boolean
      updated_at:
        type: string
    type: object
...
  api.saveFrpBody:
    properties:
      enabled:
        type: boolean
      proxies:
        items:
          $ref: '#/definitions/store.FrpTunnel'
        type: array
      server_addr:
        type: string
      server_port:
        type: integer
      token:
        type: string
    type: object
...
  store.FrpStatus:
    properties:
      error:
        type: string
      frp_version:
        type: string
      running:
        type: boolean
      ts:
        type: string
      tunnels:
        items:
          $ref: '#/definitions/store.FrpTunnelStatus'
        type: array
    type: object
  store.FrpTunnel:
    properties:
      local_ip:
        type: string
      local_port:
        type: integer
      name:
        type: string
      remote_port:
        type: integer
      type:
        description: tcp | udp | http | https | stcp ...
        type: string
    type: object
  store.FrpTunnelStatus:
    properties:
      err:
        type: string
      name:
        type: string
      rx_bytes:
        type: integer
      status:
        description: online | offline | error
        type: string
      tx_bytes:
        type: integer
      type:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: litesentry
    url: https://github.com/
  description: 轻量主机 + 容器监控 Server REST 接口（前端面板取数）。Agent↔Server 走 gRPC，契约见 docs/proto.html。
  termsOfService: http://localhost:8080/
  title: litesentry API
  version: 0.1.0
paths:
...
  /frp:
    get:
      description: 全部 frps/frpc 配置，join 最近上报状态与目标节点主机名（token 不回传）
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.frpView'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: FRP 配置列表
      tags:
      - FRP
  /frp/{kind}/{agent_id}:
    delete:
      description: 删除 frps/frpc 配置并下线对应 frp 进程（agent 下次心跳生效）
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 删除 FRP 配置
      tags:
      - FRP
    put:
      consumes:
      - application/json
      description: 新建（agent_id 不存在时）或更新 frps/frpc 配置；token 留空 = 保留原值
      parameters:
      - description: frps | frpc
        in: path
        name: kind
        required: true
        type: string
      - description: 目标 agent（frps 恒为 server）
        in: path
        name: agent_id
        required: true
        type: string
      - description: 配置体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.saveFrpBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.frpView'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 保存 FRP 配置
      tags:
      - FRP
...
securityDefinitions:
  BearerAuth:
    description: Web 面板 JWT（POST /api/auth/login 获取），格式：Bearer <jwt>
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"